handle projects that do not have package.json in root
Compares the directory containing .git and the directory of containing husky to find the directory of the package.json file.  The relative path derived from this comparison is then added to the scripts in the .git/hooks files.  

This allows npm run --json to execute properly, instead of resulting in the following error:
```
npm ERR! npm run-script [<pkg>] <command>
npm ERR! not ok code 0
```